feat: break cyclic import for ProtoV6ProviderFactories
