+++ internal/teste2e/testing.go
@@ -1,17 +1,14 @@
 package teste2e
 
 import (
-	"context"
 	"errors"
 	"os"
 	"path/filepath"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
 	"github.com/joho/godotenv"
 
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
-	tfhcloud "github.com/hetznercloud/terraform-provider-hcloud/hcloud"
 	"github.com/hetznercloud/terraform-provider-hcloud/internal/testsupport"
 )
 
@@ -46,21 +43,6 @@ var (
 	// TestLocationName is the default location where we execute our tests.
 	TestLocationName = getEnv("TEST_LOCATION", "hel1")
 )
-
-func ProtoV6ProviderFactories() map[string]func() (tfprotov6.ProviderServer, error) {
-	return map[string]func() (tfprotov6.ProviderServer, error){
-		"hcloud": func() (tfprotov6.ProviderServer, error) {
-			ctx := context.Background()
-
-			providerFactory, err := tfhcloud.GetMuxedProvider(ctx)
-			if err != nil {
-				return nil, err
-			}
-
-			return providerFactory(), nil
-		},
-	}
-}
 
 // PreCheck checks if all conditions for an acceptance test are
 // met.
